40 Player Class

+++ evade/Process.cpp
@@ -1,6 +1,6 @@
 #include "Game.h"
 
-void Process::sleep(WORD time, void (*func)(Process *me)) {
+void Process::sleep(BYTE time, void (*func)(Process *me)) {
   this->timer = time;
   if (func) {
     this->run = func;
